pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "spotipy>=2.24.0",
    "typer>=0.15.0",
    "openai>=1.55.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
voice-assistant = "src.main:app"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]
